requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
delaunay==1.0.5
Dijkstar==2.6.0
exceptiongroup==1.2.0
geojson==2.5.0
html5lib==1.1
idna==3.4
iniconfig==2.0.0
metar-taf-parser-mivek==1.8.2
packaging==24.0
pluggy==1.4.0
psycopg2==2.9.5
pytest==8.1.1
requests==2.31.0
simplekml==1.3.6
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
urllib3==1.26.15
webencodings==0.5.1
